Added char string formater.
Added the first real commit that includes the first formaters that will
be used for the printf function. At first on bonus tasks are
considered. The int_to_hex function is tested with some simple values.
Extra functionality to the int_to_hex should be added to be able to
print without prefix & use large hex notation.
The int_to_hex has to be compiled with the noted dependencies.
The compilation command is located at the last line of int_to_hex.
The command should only used for test of the int_to_hex funciton.