Make file works now
Made the explicit rules for the Makefile so that the whole project can
be compiled. Tested make all and fclean funciton.
The ft_print should still be written.
There is weird behavior that the va_arg does not accept type char.
The type is casted after in the function parameters.